Rename python file
